LayoutTests/fast/images/decode-static-image-reject.html

<head>
    <script src="../../resources/js-test.js"></script>
</head>
<body>
    <div></div>
    <script>
        description("Test rejecting the decode() promise when loading the image fails.");
        jsTestIsAsync = true;

        if (window.internals)
            internals.clearMemoryCache();

        (async () => {
            let image = new Image;

            await image.decode().catch(reason => {
                debug("Failed to decode image with no source. Result is: " + reason);
            });

            image.src = "wrongname.png";
            await image.decode().catch(reason => {
                debug("Failed to decode image with non-existent source. Result is: " + reason);
            });

            image.src = "100x100-red.psd";
            await image.decode().catch(reason => {
                debug("Failed to decode image with unsupported image format. Result is: " + reason);
            });

            image.src = "https://server:80a80/";
            await image.decode().catch(reason => {
                debug("Failed to decode image with invalid URL. Result is: " + reason);
            });

            finishJSTest();
        })();
    </script>
</body>
