LayoutTests/fast/images/decode-static-image-reject-expected.txt

Test rejecting the decode() promise when loading the image fails.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Failed to decode image with no source. Result is: EncodingError: Missing source URL.
Failed to decode image with non-existent source. Result is: EncodingError: Loading error.
Failed to decode image with unsupported image format. Result is: EncodingError: Loading error.
Failed to decode image with invalid URL. Result is: EncodingError: Loading error.
PASS successfullyParsed is true

TEST COMPLETE
